+++ CHANGELOG.md
@@ -52,6 +52,7 @@
 - fixed number of output-values in case of int and float output-type
 - task-abort-endpoint not works
 - task-abort-modal was fixed in the dashboard
+- fixed delete model with running tasks
 
 ## v0.10.0